refactor: allow robots and animals to have AE that change derived data
