+++ src/module/hooks/diceTrayIntegration.ts
@@ -1,7 +1,7 @@
 // DICE TRAY no longer supports custom buttons through a simple API as of V12
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-nocheck This turns off *all* typechecking, make sure to remove this once foundry-vtt-types are updated to cover v10.
-/*import { getKeyByValue } from "../utils/sheetUtils";
+/*import { getKeyByValue } from "../utils/utils";
 import { simplifySkillName } from "../utils/utils";
 import {TWODSIX} from "../config";
