+++ src/module/utils/TwodsixDiceRoll.ts
@@ -5,7 +5,7 @@ import {TWODSIX} from "../config";
 import TwodsixActor from "../entities/TwodsixActor";
 import TwodsixItem from "../entities/TwodsixItem";
 import {advantageDisadvantageTerm} from "../i18n";
-import {getKeyByValue} from "./sheetUtils";
+import {getKeyByValue} from "./utils";
 import {TwodsixRollSettings} from "./TwodsixRollSettings";
 import Crit from "./crit";
 import { simplifySkillName, addSign, capitalizeFirstLetter } from "./utils";
